+++ src/py/pyodide/_core.py
@@ -1,9 +1,11 @@
 import sys
 
+IN_BROWSER = True
 if "_pyodide_core" not in sys.modules:
     from _pyodide import _core_docs as _pyodide_core
 
     sys.modules["_pyodide_core"] = _pyodide_core
+    IN_BROWSER = False
 
 from _pyodide_core import (
     JsProxy,
